Move constants from the SDK to common
