+++ rock/common/constants.py
@@ -1,1 +1,3 @@
 GET_STATUS_SWITCH = "get_status_v2_enabled"
+PID_PREFIX = "PIDSTART"
+PID_SUFFIX = "PIDEND"
