+++ rock/sdk/common/constants.py
@@ -19,5 +19,3 @@ class Constants(metaclass=DeprecatedMeta):
 
 
 RunModeType = Literal["normal", "nohup"]
-PID_PREFIX = "PIDSTART"
-PID_SUFFIX = "PIDEND"
